+++ src/test/java/com/notebookmanager/controller/AlunoControllerSuccessTest.java
@@ -21,10 +21,10 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class AlunoControllerTest {
+public class AlunoControllerSuccessTest {
 
     @Autowired
-    TestRestTemplate restTemplate;
+    private TestRestTemplate restTemplate;
 
 
     @Test
@@ -46,15 +46,6 @@ public class AlunoControllerTest {
         assertThat(curso).isEqualTo("MEDICINA");
     }
 
-    @Test
-    void naoRetornaAlunoComRaInvalido() {
-        ResponseEntity<String> response = restTemplate.getForEntity("/alunos/96", String.class);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
-
-        assertThat(response.getBody()).isBlank();
-
-    }
 
     @Test
     @DirtiesContext
@@ -80,22 +71,7 @@ public class AlunoControllerTest {
 
     }
 
-    @Test
-    void naoSalvaAlunoComRaRepetido() {
-        Aluno aluno = new Aluno("Jonathan Luciano", "09135616", "jonathanluciano@puccampinas.edu.br", "(19)94291-7013",
-                Curso.NUTRICAO, LocalDateTime.now(), LocalDateTime.now());
 
-        ResponseEntity<String> response = restTemplate.postForEntity("/alunos", aluno, String.class);
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CONFLICT);
-
-        DocumentContext documentContext = JsonPath.parse(response.getBody());
-
-        String status =  documentContext.read("$.status");
-        assertThat(status).isEqualTo("CONFLICT");
-
-        String message = documentContext.read("$.message");
-        assertThat(message).isEqualTo("O Aluno com este RA já está cadastrado.");
-    }
 
     @Test
     void retornaListaDeAlunos() {
@@ -197,26 +173,6 @@ public class AlunoControllerTest {
         assertThat(atualizadoEm).isNotEqualTo("2012-11-10T21:12:37");
     }
 
-    @Test
-    void naoAtualizaAlunoNaoExistente() {
-        Aluno aluno = new Aluno("Josue Nao Existe No Banco", "02020202", "josuenenb@puccampinas.edu.br", "(19)93123-4231",
-                Curso.ODONTOLOGIA, LocalDateTime.of(2012, 11, 10, 21, 12, 37),
-                LocalDateTime.of(2012, 11, 10, 21, 12, 37));
-        HttpEntity<Aluno> request = new HttpEntity<>(aluno);
-
-        ResponseEntity<String> response = restTemplate.exchange("/alunos/90", HttpMethod.PUT,
-                request, String.class);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
-
-        DocumentContext documentContext =  JsonPath.parse(response.getBody());
-
-        String status =  documentContext.read("$.status");
-        assertThat(status).isEqualTo("NOT_FOUND");
-
-        String message = documentContext.read("$.message");
-        assertThat(message).isEqualTo("Aluno não encontrado.");
-    }
 
     @Test
     @DirtiesContext
@@ -232,19 +188,6 @@ public class AlunoControllerTest {
         assertThat(getResponse.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
     }
 
-    @Test
-    void naoDeleteAlunoNaoExistente() {
-        ResponseEntity<String> response = restTemplate.exchange("/alunos/31", HttpMethod.DELETE,
-                null, String.class);
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
 
-        DocumentContext documentContext =  JsonPath.parse(response.getBody());
-
-        String status =  documentContext.read("$.status");
-        assertThat(status).isEqualTo("NOT_FOUND");
-
-        String message = documentContext.read("$.message");
-        assertThat(message).isEqualTo("Aluno não encontrado.");
-    }
 
 }
